+++ src/Components/TopBar.tsx
@@ -7,7 +7,7 @@ export const TopBar = () => {
     const MoonIcon = 'assets/icons/moon.svg';
 
     return (
-        <div className={`flex justify-between p-5 shadow-sm items-center ${theme}`}>
+        <div data-testid="top-bar" className={`flex justify-between p-5 shadow-sm items-center ${theme}`}>
             <p className='font-extrabold text-xl'>Where in the world?</p>
             <div className='flex' onClick={() => changeTheme(theme)}>
                 <img src={MoonIcon} alt="moon" className='w-5 mr-2 bg-white cursor-pointer' />